Address JSON-P and JSON-B review feedback
