+++ config/checkstyle/suppressions.xml
@@ -9,5 +9,4 @@
               <!-- files="DefaultBeanContext.java|BeanDefinitionWriter.java|DefaultHttpClient.java"/> -->
 
     <suppress checks="MissingJavadocType" files=".*doc-examples.*" />
-    <suppress checks="FileLength" files="MicronautJsonProvider.java" />
 </suppressions>
